feat: ProjectDetail.vue -> Project.vue 파일명 변경
